+++ DesktopApplicationTemplate.UI/Views/HttpServiceView.xaml.cs
@@ -16,6 +16,7 @@ using DesktopApplicationTemplate.UI.Services;
 
 using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.ViewModels;
+using DesktopApplicationTemplate.Models;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace DesktopApplicationTemplate.UI.Views
@@ -45,7 +46,8 @@ namespace DesktopApplicationTemplate.UI.Views
 
         public void SetServiceContext(ServiceListModel service)
         {
-            LogView.DataContext = new ServiceLogViewModel(service.DisplayName, service.ServiceType, service.Logs);
+            Enum.TryParse<ServiceType>(service.ServiceType, true, out var type);
+            LogView.DataContext = new ServiceLogViewModel(type, service.Logs);
         }
 
         private void Help_Click(object sender, RoutedEventArgs e)